requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-groq>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
groq>=0.4.0
streamlit>=1.28.0
python-dotenv>=1.0.0